register eye password view
