+++ proptech/src/App.jsx
@@ -33,7 +33,7 @@ import AdministrationDashboard from "./components/AdministrationDashboard";
 import CapitalizationCalculatorAdmin from "./components/CapitalizationCalculatorAdmin";
 import CapitalizationCalculatorEdit from "./components/CapitalizationCalculatorEdit";
 import Dashboard from "./components/Dashboard";
-import RegisterUser from "./components/RegisterUser";
+import RegisterUserAdmin from "./components/RegisterUserAdmin";
 
 function App() {
     return (
@@ -68,7 +68,7 @@ function App() {
                 <Route path="/payment" element={<PaymentForm/>} />
                 <Route path="/adminDashboard" element={<AdministrationDashboard/>} />
                 <Route path="/dashboard" element={<Dashboard />} />
-                <Route path="/registerUser" element={<RegisterUser/>} />
+                <Route path="/registerUserAdmin" element={<RegisterUserAdmin/>} />
             </Route>
             <Route path="/" element={<Home />} />
             <Route path="*" element={<ErrorPage />} />
