+++ proptech/src/context/Context.jsx
@@ -101,6 +101,37 @@ export const ContextProvider = ({ children }) => {
     localStorage.clear();
     navigate("/");
   };
+//for register new user from Admin dashboard
+  const registerUserAdmin = async (data) => {
+    try {
+      const url =
+        "https://h3-20-proptech-production.up.railway.app/api/register/";
+
+      const response = await fetch(url, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          Accept: "application/json",
+        },
+        body: JSON.stringify(data),
+        credentials: "include",
+      });
+      console.log("Before SHOW:::", response);
+
+      if (!response.ok) {
+        throw new Error(`HTTP error! status: ${response.status}`);
+      }
+
+      const result = await response.json();
+      console.log("Registration successful:", result);
+    //   navigate("/login");
+      return result;
+    } catch (error) {
+      console.error("Registration error:", error.message);
+      throw error;
+    }
+  };
+
 
   return (
     <Context.Provider
@@ -113,6 +144,7 @@ export const ContextProvider = ({ children }) => {
         loginUser,
         logoutUser,
         registerUser,
+        registerUserAdmin,
         isAuthenticated,
       }}
     >
